Fixed bug

+++ src/CreateGraph.cu
@@ -55,7 +55,7 @@ __global__ void createCornerGraphKernel(unsigned char *image, unsigned int *d_ve
             other_r = image[other_img_idx];
             other_g = image[other_img_idx + 1];
             other_b = image[other_img_idx + 2];
-            distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+            distance = 8.0*sqrtf(powf((this_r - other_r), 2) + powf((this_g - other_g), 2) + powf((this_b - other_b), 2.0));
             d_weight[write_offset] = (unsigned int) round(distance);
         }
 
@@ -67,7 +67,7 @@ __global__ void createCornerGraphKernel(unsigned char *image, unsigned int *d_ve
             other_r = image[other_img_idx];
             other_g = image[other_img_idx + 1];
             other_b = image[other_img_idx + 2];
-            distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+            distance = 8.0*sqrtf(powf((this_r - other_r), 2) + powf((this_g - other_g), 2) + powf((this_b - other_b), 2.0));
             d_weight[write_offset] = (unsigned int) round(distance);
         }
 
@@ -79,7 +79,7 @@ __global__ void createCornerGraphKernel(unsigned char *image, unsigned int *d_ve
             other_r = image[other_img_idx];
             other_g = image[other_img_idx + 1];
             other_b = image[other_img_idx + 2];
-            distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+            distance = 8.0*sqrtf(powf((this_r - other_r), 2.0) + powf((this_g - other_g), 2.0) + powf((this_b - other_b), 2.0));
             d_weight[write_offset+1] = (unsigned int) round(distance);
         }
 
@@ -91,7 +91,7 @@ __global__ void createCornerGraphKernel(unsigned char *image, unsigned int *d_ve
             other_r = image[other_img_idx];
             other_g = image[other_img_idx + 1];
             other_b = image[other_img_idx + 2];
-            distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+            distance = 8.0*sqrtf(powf((this_r - other_r), 2.0) + powf((this_g - other_g), 2.0) + powf((this_b - other_b), 2.0));
             d_weight[write_offset+1] = (unsigned int) round(distance);
         }
     }
@@ -131,7 +131,7 @@ __global__ void createFirstRowGraphKernel(unsigned char *image, unsigned int *d_
         other_r = image[other_img_idx];
         other_g = image[other_img_idx + 1];
         other_b = image[other_img_idx + 2];
-        distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+        distance = 8.0*sqrtf(powf((this_r - other_r), 2.0) + powf((this_g - other_g), 2.0) + powf((this_b - other_b), 2.0));
         d_weight[write_offset] = (unsigned int) round(distance);
 
         // Right node
@@ -141,7 +141,7 @@ __global__ void createFirstRowGraphKernel(unsigned char *image, unsigned int *d_
         other_r = image[other_img_idx];
         other_g = image[other_img_idx + 1];
         other_b = image[other_img_idx + 2];
-        distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+        distance = 8.0*sqrtf(powf((this_r - other_r), 2.0) + powf((this_g - other_g), 2.0) + powf((this_b - other_b), 2.0));
         d_weight[write_offset+1] = (unsigned int) round(distance);
 
         // Bottom node
@@ -151,7 +151,7 @@ __global__ void createFirstRowGraphKernel(unsigned char *image, unsigned int *d_
         other_r = image[other_img_idx];
         other_g = image[other_img_idx + 1];
         other_b = image[other_img_idx + 2];
-        distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+        distance = 8.0*sqrtf(powf((this_r - other_r), 2.0) + powf((this_g - other_g), 2.0) + powf((this_b - other_b), 2.0));
         d_weight[write_offset+2] = (unsigned int) round(distance);
     }
 }
@@ -192,7 +192,7 @@ __global__ void createLastRowGraphKernel(unsigned char *image, unsigned int *d_v
         other_r = image[other_img_idx];
         other_g = image[other_img_idx + 1];
         other_b = image[other_img_idx + 2];
-        distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+        distance = 8.0*sqrtf(powf((this_r - other_r), 2.0) + powf((this_g - other_g), 2.0) + powf((this_b - other_b), 2.0));
         d_weight[write_offset] = (unsigned int) round(distance);
 
         // Right node
@@ -202,7 +202,7 @@ __global__ void createLastRowGraphKernel(unsigned char *image, unsigned int *d_v
         other_r = image[other_img_idx];
         other_g = image[other_img_idx + 1];
         other_b = image[other_img_idx + 2];
-        distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+        distance = 8.0*sqrtf(powf((this_r - other_r), 2.0) + powf((this_g - other_g), 2.0) + powf((this_b - other_b), 2.0));
         d_weight[write_offset+1] = (unsigned int) round(distance);
 
         // Top node
@@ -212,7 +212,7 @@ __global__ void createLastRowGraphKernel(unsigned char *image, unsigned int *d_v
         other_r = image[other_img_idx];
         other_g = image[other_img_idx + 1];
         other_b = image[other_img_idx + 2];
-        distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+        distance = 8.0*sqrtf(powf((this_r - other_r), 2.0) + powf((this_g - other_g), 2.0) + powf((this_b - other_b), 2.0));
         d_weight[write_offset+2] = (unsigned int) round(distance);
     }
 }
@@ -252,7 +252,7 @@ __global__ void createFirstColumnGraphKernel(unsigned char *image, unsigned int 
         other_r = image[other_img_idx];
         other_g = image[other_img_idx + 1];
         other_b = image[other_img_idx + 2];
-        distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+        distance = 8.0*sqrtf(powf((this_r - other_r), 2.0) + powf((this_g - other_g), 2.0) + powf((this_b - other_b), 2.0));
         d_weight[write_offset] = (unsigned int) round(distance);
 
         // Bottom node
@@ -262,7 +262,7 @@ __global__ void createFirstColumnGraphKernel(unsigned char *image, unsigned int 
         other_r = image[other_img_idx];
         other_g = image[other_img_idx + 1];
         other_b = image[other_img_idx + 2];
-        distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+        distance = 8.0*sqrtf(powf((this_r - other_r), 2.0) + powf((this_g - other_g), 2.0) + powf((this_b - other_b), 2.0));
         d_weight[write_offset+1] = (unsigned int) round(distance);
 
         // Top node
@@ -272,7 +272,7 @@ __global__ void createFirstColumnGraphKernel(unsigned char *image, unsigned int 
         other_r = image[other_img_idx];
         other_g = image[other_img_idx + 1];
         other_b = image[other_img_idx + 2];
-        distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+        distance = 8.0*sqrtf(powf((this_r - other_r), 2.0) + powf((this_g - other_g), 2.0) + powf((this_b - other_b), 2.0));
         d_weight[write_offset+2] = (unsigned int) round(distance);
     }
 }
@@ -313,7 +313,7 @@ __global__ void createLastColumnGraphKernel(unsigned char *image, unsigned int *
         other_r = image[other_img_idx];
         other_g = image[other_img_idx + 1];
         other_b = image[other_img_idx + 2];
-        distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+        distance = 8.0*sqrtf(powf((this_r - other_r), 2.0) + powf((this_g - other_g), 2.0) + powf((this_b - other_b), 2.0));
         d_weight[write_offset] = (unsigned int) round(distance);
 
         // Bottom node
@@ -323,7 +323,7 @@ __global__ void createLastColumnGraphKernel(unsigned char *image, unsigned int *
         other_r = image[other_img_idx];
         other_g = image[other_img_idx + 1];
         other_b = image[other_img_idx + 2];
-        distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+        distance = 8.0*sqrtf(powf((this_r - other_r), 2.0) + powf((this_g - other_g), 2.0) + powf((this_b - other_b), 2.0));
         d_weight[write_offset+1] = (unsigned int) round(distance);
 
         // Top node
@@ -333,7 +333,7 @@ __global__ void createLastColumnGraphKernel(unsigned char *image, unsigned int *
         other_r = image[other_img_idx];
         other_g = image[other_img_idx + 1];
         other_b = image[other_img_idx + 2];
-        distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+        distance = 8.0*sqrtf(powf((this_r - other_r), 2.0) + powf((this_g - other_g), 2.0) + powf((this_b - other_b), 2.0));
         d_weight[write_offset+2] = (unsigned int) round(distance);
     }
 }
@@ -375,7 +375,7 @@ __global__ void createInnerGraphKernel(unsigned char *image, unsigned int *d_ver
         other_r = image[other_img_idx];
         other_g = image[other_img_idx + 1];
         other_b = image[other_img_idx + 2];
-        distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+        distance = 8.0*sqrtf(powf((this_r - other_r), 2.0) + powf((this_g - other_g), 2.0) + powf((this_b - other_b), 2.0));
         d_weight[write_offset] = (unsigned int) round(distance);
 
         // Right node
@@ -385,7 +385,7 @@ __global__ void createInnerGraphKernel(unsigned char *image, unsigned int *d_ver
         other_r = image[other_img_idx];
         other_g = image[other_img_idx + 1];
         other_b = image[other_img_idx + 2];
-        distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+        distance = 8.0*sqrtf(powf((this_r - other_r), 2.0) + powf((this_g - other_g), 2.0) + powf((this_b - other_b), 2.0));
         d_weight[write_offset+1] = (unsigned int) round(distance);
 
         // Bottom node
@@ -395,7 +395,7 @@ __global__ void createInnerGraphKernel(unsigned char *image, unsigned int *d_ver
         other_r = image[other_img_idx];
         other_g = image[other_img_idx + 1];
         other_b = image[other_img_idx + 2];
-        distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+        distance = 8.0*sqrtf(powf((this_r - other_r), 2.0) + powf((this_g - other_g), 2.0) + powf((this_b - other_b), 2.0));
         d_weight[write_offset+2] = (unsigned int) round(distance);
 
         // Top node
@@ -405,7 +405,7 @@ __global__ void createInnerGraphKernel(unsigned char *image, unsigned int *d_ver
         other_r = image[other_img_idx];
         other_g = image[other_img_idx + 1];
         other_b = image[other_img_idx + 2];
-        distance = 8.0*sqrt(pow((this_r - other_r), 2) + pow((this_g - other_g), 2) + pow((this_b - other_b), 2));
+        distance = 8.0*sqrtf(powf((this_r - other_r), 2.0) + powf((this_g - other_g), 2.0) + powf((this_b - other_b), 2.0));
         d_weight[write_offset+3] = (unsigned int) round(distance);
     }
 }
